+++ stroom-core-server/src/main/java/stroom/proxy/repo/ZipFragmenter.java
@@ -29,7 +29,6 @@ import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.Set;
 
 class ZipFragmenter {
@@ -41,18 +40,14 @@ class ZipFragmenter {
         this.errorReceiver = errorReceiver;
     }
 
-    public void fragment(final Path path, final BasicFileAttributes attrs) {
+    public void fragment(final Path path) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("Getting zip info for  '" + FileUtil.getCanonicalPath(path) + "'");
         }
 
         // Create output dir.
-        final String fileName = path.getFileName().toString();
-        final int index = fileName.lastIndexOf(".");
-        if (index != -1) {
-            final String stem = fileName.substring(0, index);
-            final Path outputDir = path.getParent().resolve(stem + PathConstants.PARTS);
-
+        final Path outputDir = PartsPathUtil.createPartsDir(path);
+        if (outputDir != null) {
             if (!Files.isDirectory(outputDir)) {
                 try {
                     Files.createDirectory(outputDir);
@@ -92,7 +87,7 @@ class ZipFragmenter {
                                 }
                             }
 
-                            final Path outputFile = currentDir.resolve(stem + PathConstants.PART + idString + ".zip");
+                            final Path outputFile = PartsPathUtil.createPart(currentDir, path, idString);
                             // If output file already exists then it ought to be overwritten automatically.
                             try (final StroomZipOutputStream stroomZipOutputStream = new StroomZipOutputStreamImpl(outputFile)) {
                                 transferEntry(stroomZipFile, stroomZipOutputStream, baseName, StroomZipFileType.Meta);
@@ -113,7 +108,7 @@ class ZipFragmenter {
                 if (moveOriginalFile) {
                     try {
                         final String idString = StroomFileNameUtil.idToString(i);
-                        final Path outputFile = currentDir.resolve(stem + PathConstants.PART + idString + ".zip");
+                        final Path outputFile = PartsPathUtil.createPart(currentDir, path, idString);
                         Files.move(
                                 path,
                                 outputFile,